ClientSide/ClientServer/requirements.txt
google-api-python-client
grpcio
grpcio-tools
psutil